Smaller compression-related fixes
Also update spirv-simulator to latest version

+++ src/util.h
@@ -83,7 +83,7 @@ extern uint_fast8_t p__disable_multithread_read;
 extern uint_fast8_t p__allow_stalls;
 extern uint_fast16_t p__preload;
 extern uint_fast8_t p__compression_type;
-extern uint_fast8_t p__compression_level;
+extern uint_fast16_t p__compression_level;
 
 /// Logging to be enable as needed by source recompilation
 #define NEVER(_format, ...)
